move unformat to utils.js
